Reorganizing tests packages
