+++ android-test/app/src/test/java/co/elastic/apm/android/test/attributes/traces/GlobalAttributeTest.java
@@ -1,4 +1,4 @@
-package co.elastic.apm.android.test.spanattributes;
+package co.elastic.apm.android.test.attributes.traces;
 
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.verify;
@@ -9,6 +9,7 @@ import java.util.List;
 
 import co.elastic.apm.android.sdk.internal.time.ntp.NtpManager;
 import co.elastic.apm.android.sdk.traces.common.tools.ElasticTracer;
+import co.elastic.apm.android.test.attributes.traces.SpanAttrHost;
 import co.elastic.apm.android.test.common.spans.Spans;
 import co.elastic.apm.android.test.testutils.TestElasticClock;
 import co.elastic.apm.android.test.testutils.base.BaseRobolectricTest;
