+++ android-test/app/src/test/java/co/elastic/apm/android/test/attributes/traces/ResourcesTest.java
@@ -1,4 +1,4 @@
-package co.elastic.apm.android.test.spanattributes;
+package co.elastic.apm.android.test.attributes.traces;
 
 import android.os.Build;
